HSEARCH-3892 Move type-checking and type-conversion tests for single-field aggregations to a dedicated class
